Move ArchiveHelper into common module
