+++ airbyte-commons/build.gradle
@@ -4,4 +4,6 @@ plugins {
 
 dependencies {
     testImplementation 'org.apache.commons:commons-lang3:3.11'
+
+    implementation group: 'org.apache.commons', name: 'commons-compress', version: '1.20'
 }
